Move unit tests to __tests__ folders
