+++ src/Tracker/__mocks__/safeWoopra.js
@@ -1,0 +1,11 @@
+/* eslint-disable @typescript-eslint/no-empty-function, @typescript-eslint/explicit-module-boundary-types */
+// @TODO: Put proper implementation to test Tracker
+class WoopraTracker {
+  init() {}
+  config() {}
+  identify() {}
+  track() {}
+  dispose() {}
+}
+
+export default WoopraTracker
