+++ src/components/App/__tests__/index.test.js
@@ -1,7 +1,7 @@
 import { h } from 'preact'
 import { shallow } from 'enzyme'
 
-import App from './index'
+import App from '../index'
 
 it('renders without crashing', () => {
   shallow(<App />)
