+++ src/demo/__tests__/demo.test.tsx
@@ -7,7 +7,9 @@ declare global {
   }
 }
 
-jest.mock('./demoUtils', () => ({
+jest.mock('../../Tracker/safeWoopra')
+
+jest.mock('../demoUtils', () => ({
   getInitSdkOptions: jest.fn().mockReturnValue({}),
   queryParamToValueString: { useHistory: false },
   getTokenFactoryUrl: () => 'https://token-factory.onfido.com/sdk_token',
@@ -18,17 +20,8 @@ jest.mock('./demoUtils', () => ({
     ),
 }))
 
-// when the Onfido SDK is imported Woopra needs to be mocked
-const mockWoopraFn = jest.fn()
-
-jest.mock('../Tracker/safeWoopra', () =>
-  jest.fn().mockImplementation(() => ({
-    init: () => mockWoopraFn,
-    config: () => mockWoopraFn,
-    identify: () => mockWoopraFn,
-    track: () => mockWoopraFn,
-  }))
-)
+const mockedConsole = jest.fn()
+console.log = mockedConsole
 
 describe('Mount Demo App', () => {
   let Demo: FunctionComponent = null
@@ -41,6 +34,10 @@ describe('Mount Demo App', () => {
     document.body.appendChild(rootNode)
   })
 
+  afterEach(() => {
+    jest.clearAllMocks()
+  })
+
   describe('by mocking Onfido SDK', () => {
     beforeEach(() => {
       window.Onfido = {
@@ -51,24 +48,40 @@ describe('Mount Demo App', () => {
         })),
       }
 
-      Demo = require('./demo').Demo
+      Demo = require('../demo').Demo
     })
 
     it('mounts the Onfido Demo without crashing', () => {
       const sdkDemo = mount(<Demo />)
       expect(sdkDemo.exists()).toBeTruthy()
       expect(window.Onfido.init).toHaveBeenCalled()
+      expect(mockedConsole).toHaveBeenCalledWith(
+        '* JWT Factory URL:',
+        'https://token-factory.onfido.com/sdk_token',
+        'for',
+        'EU',
+        'in',
+        'test'
+      )
     })
   })
 
   describe('without mocking Onfido SDK', () => {
     beforeEach(() => {
-      Demo = require('./demo').Demo
+      Demo = require('../demo').Demo
     })
 
     it('mounts the Onfido Demo without crashing', () => {
       const sdkDemo = mount(<Demo />)
       expect(sdkDemo.exists()).toBeTruthy()
+      expect(mockedConsole).toHaveBeenCalledWith(
+        '* JWT Factory URL:',
+        'https://token-factory.onfido.com/sdk_token',
+        'for',
+        'EU',
+        'in',
+        'test'
+      )
     })
   })
 })
