+++ src/supported-documents/__tests__/supported-documents.test.js
@@ -1,4 +1,4 @@
-import { getSupportedCountriesForDocument } from './index'
+import { getSupportedCountriesForDocument } from '../index'
 
 describe('getSupportedCountriesForDocument', () => {
   it('should not contain any duplicate countries', () => {
